pyproject.toml
[project]
name = "bigquery-append"
version = "0.1.0"
requires-python = ">=3.9"
dependencies = [
    "flask==3.0.0",
    "flask-sqlalchemy>=3.1.1",
    "google-api-python-client==2.118.0",
    "google-auth==2.27.0",
    "google-auth-httplib2==0.2.0",
    "google-auth-oauthlib==1.2.0",
    "google-cloud-bigquery==3.14.0",
    "markdown>=3.8.2",
    "pandas==2.2.0",
    "pyarrow>=20.0.0",
]

[tool.uv]
dev-dependencies = [] 
